+++ packages/iov-core/src/multichainsigner.spec.ts
@@ -1,13 +1,13 @@
 import {
   Address,
   Algorithm,
-  BcpTransactionState,
   ChainId,
   isSendTransaction,
   PublicIdentity,
   PublicKeyBytes,
   SendTransaction,
   TokenTicker,
+  TransactionState,
 } from "@iov/bcp-types";
 import { bnsCodec, bnsConnector } from "@iov/bns";
 import { Ed25519, Random } from "@iov/crypto";
@@ -103,7 +103,7 @@ describe("MultiChainSigner", () => {
         },
       };
       const postResponse = await signer.signAndPost(sendTx, mainWalletId);
-      await postResponse.blockInfo.waitFor(info => info.state === BcpTransactionState.InBlock);
+      await postResponse.blockInfo.waitFor(info => info.state !== TransactionState.Pending);
 
       // we should be a little bit richer
       const updatedAccount = await connection.getAccount({ address: recipient });