+++ CHANGELOG.md
@@ -35,6 +35,10 @@ Breaking changes
 * @iov/bcp-types: Migrate `UnsignedTransaction.chainId` and `.signer` into `.creator`
 * @iov/bcp-types: Add `BcpConnection.getNonces`
 * @iov/bcp-types: Add `BcpTxQuery.sentFromOrTo` in favour of package-specific address tags
+* @iov/bcp-types: Removed `Bcp` prefix from `BcpTransactionState`, `BcpBlockInfoPending`,
+  `BcpBlockInfoInBlock`, `BcpBlockInfo`.
+* @iov/bcp-types: Add block info state `BlockInfoFailed`.
+  `BcpBlockInfoInBlock`, `BcpBlockInfo`.
 * @iov/bns: Convert `.owner` in `BnsUsernameNft` and `BnsBlockchainNft` to `Address`
 * @iov/bns: In `BncConnection.watchBlockHeaders`, the block header events
   temporarily contain a dummy `id` string until ID calculation is implemented